move BIP85 class to __init__.py to avoid conflict with the file
Since there is both a folder/package named bip85 and a file, importing
it can cause confusion and issues. For example `from bip85 import app`
can fail because Python may try to import it from `bip85/bip85.py`
instead of `bip85/app.py`, depending on `sys.path`, `PYTHONPATH`, etc.
